# docs/swagger.yaml
basePath: /
definitions:
  dto.HealthzResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  dto.RabbitSendRequest:
    properties:
      message:
        example: hello rabbit
        type: string
      source:
        example: go-service
        type: string
    required:
    - message
    - source
    type: object
  dto.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Go Service API
  title: Go Service API
  version: "1.0"
paths:
  /healthz:
    get:
      description: Returns service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthzResponse'
      summary: Health check
      tags:
      - health
  /v1/example/plus/{ID}:
    post:
      description: Находит в БД счётчик с указанным ID, создаёт в случае отсутствия,
        и увеличивает значение счётчика на указанное значение
      operationId: plus-value
      parameters:
      - description: Counter ID
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Увеличивает значение счётчика в БД
      tags:
      - example
  /v1/rabbit/send:
    post:
      consumes:
      - application/json
      description: Accepts JSON data and sends it to RabbitMQ queue
      operationId: rabbit-send
      parameters:
      - description: RabbitMQ JSON message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RabbitSendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Send JSON message to RabbitMQ queue
      tags:
      - rabbitmq
swagger: "2.0"